Refactor 폴더 구조 수정 및  커스텀 훅 제작
