+++ src/App.js
@@ -1,15 +1,15 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import PersonalFolder from "./components/Folder/SharedPage";
-import NewPage from "./components/NewPage/index";
-import FolderPage from "./components/Folder/FolderPage";
+import SharedPage from "./components/Pages/SharedPage/index";
+import NewPage from "./components/Pages/NewPage/index";
+import FolderPage from "./components/Pages/FolderPage/index";
 
 function App() {
   return (
     <BrowserRouter>
       <div className="App">
         <Routes>
-          <Route path="/shared" element={<PersonalFolder />} />
+          <Route path="/shared" element={<SharedPage />} />
           <Route path="/newPage" element={<NewPage />} />
           <Route path="/folder" element={<FolderPage />} />
         </Routes>
