+++ src/components/Layout/Footer/index.jsx
@@ -1,5 +1,5 @@
 /** @jsxImportSource @emotion/react */
-import { css } from "@emotion/react";
+import styles from "./styles";
 import { NavLink } from "react-router-dom";
 import {
   FacebookIcon,
@@ -18,21 +18,21 @@ const Footer = () => {
   ];
 
   return (
-    <div css={divStyle}>
-      <section css={pStyle}>
+    <div css={styles.div}>
+      <section css={styles.p}>
         <p>@codeit - 2023</p>
       </section>
 
-      <nav css={aWrapperStyle}>
-        <NavLink to="/policy" css={aStyle}>
+      <nav css={styles.aWrapper}>
+        <NavLink to="/policy" css={styles.a}>
           Privacy Policy
         </NavLink>
-        <NavLink to="/faq" css={aStyle}>
+        <NavLink to="/faq" css={styles.a}>
           FAQ
         </NavLink>
       </nav>
 
-      <section css={iconsStyle}>
+      <section css={styles.icons}>
         {icons.map((icon) => (
           <icon.component key={icon.id} />
         ))}
@@ -41,58 +41,3 @@ const Footer = () => {
   );
 };
 export default Footer;
-
-const divStyle = css`
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  height: 160px;
-  padding: 32px 64px;
-  background: var(--Linkbrary-black);
-  margin: 0;
-`;
-const pStyle = css`
-  order: 1;
-  margin: 0;
-  color: #676767;
-  font-family: Arial;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-
-  @media screen and (375px <= width <= 767px) {
-    order: 3;
-    width: 100%;
-    margin-top: 60px;
-  }
-`;
-const aWrapperStyle = css`
-  order: 2;
-  display: flex;
-  gap: 30px;
-
-  @media screen and (375px <= width <= 767px) {
-    order: 1;
-  }
-`;
-const aStyle = css`
-  margin: 0;
-  text-decoration: none;
-  color: #cfcfcf;
-  font-family: Arial;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-`;
-const iconsStyle = css`
-  order: 3;
-  display: flex;
-  gap: 12px;
-  margin: 0;
-
-  @media screen and (375px <= width <= 767px) {
-    order: 2;
-  }
-`;
