--- src/components/NewPage/index.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const NewPage = () => {
-  return <p>새 페이지 </p>;
-};
-
-export default NewPage;
