+++ src/components/Pages/FolderPage/AddLink/styles.js
@@ -1,0 +1,20 @@
+import { css } from "@emotion/react";
+
+export const divCss = css`
+  display: flex;
+  justify-content: center;
+  height: 216px;
+`;
+export const imgCss = css`
+  position: relative;
+  left: 32px;
+  top: 4px;
+`;
+export const inputCss = css`
+  border: solid 4px var(--Linkbrary-gradation);
+  outline: 4px solid var(--Linkbrary-gradation);
+  width: 800px;
+  height: 68px;
+  border-radius: 16px;
+  padding: 16px 40px;
+`;
