+++ src/components/Pages/FolderPage/NavButton/styles.js
@@ -1,0 +1,11 @@
+import { css } from "@emotion/react";
+export const btnCss = css`
+  font-size: 16px;
+  margin-left: 12px;
+  background: var(--Linkbrary-white);
+  border: solid 1px var(--Linkbrary-primary-color);
+  border-radius: 4px;
+  padding: 8px 12px;
+`;
+
+
