+++ src/components/Pages/SharedPage/styles.js
@@ -1,0 +1,6 @@
+import { css } from "@emotion/react";
+
+export const divCss = css`
+  background: var(--Linkbrary-white);
+  padding: 40px;
+`;
